mgr/cephadm: move _check_for_strays to serve.py
